Standardise and test bundle_js_app too
