+++ python/setup.py
@@ -44,7 +44,7 @@ setup(
         path.join(PACKAGE_NAME, "split_ledger.py"),
         path.join(PACKAGE_NAME, "migrate_1_x_config.py"),
         path.join(PACKAGE_NAME, "ballot_builder.py"),
-        path.join(PACKAGE_NAME, "bundle_js_modules.py"),
+        path.join(PACKAGE_NAME, "bundle_js_app.py"),
         path.join(UTILITIES_PATH, "build_proposal.sh"),
         path.join(UTILITIES_PATH, "keygenerator.sh"),
         path.join(UTILITIES_PATH, "scurl.sh"),
